perf(featured): code-split planner out of the featured-page bundle
Lazy-load App and the featured pages in main.jsx so the planner and its routing
engine (RouteManager, base-routing decode) no longer ship in the entry chunk.
The splash now hands off on actual route mount (RouteReady + null Suspense
fallback) instead of a timer, so it stays visible during the route chunk fetch
with no blank flash.

Extract emptyRouteSnapshot into a dependency-free routeSnapshot.js (re-exported
from routeActions.js) so the featured snapshot loader stops pulling the routing
chunk for a 10-line factory.

Result: entry bundle 397KB -> 215KB; planner code (App ~139KB incl. routing)
loads only on /, not on public featured pages.
